requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# CLIP
clip-by-openai

# GUI
customtkinter>=5.0.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
